feat(product tours): support auto-launch field, reduce bundle size
