+++ .changeset/famous-loops-deny.md
@@ -1,0 +1,5 @@
+---
+'posthog-js': minor
+---
+
+product tours: support auto-show config, add modal steps
